refactor(portal): move documents view-model to lib
